Added PostgreCastBase
